Removed the sym-linked structure to behave more like the Gentoo generator (see ros-infrastructure/superflore#51).

+++ superflore/generators/bitbake/run.py
@@ -13,8 +13,6 @@
 # limitations under the License.
 
 import argparse
-import errno
-import os
 import shutil
 import sys
 
@@ -23,7 +21,7 @@ from rosinstall_generator.distro import get_distro
 from superflore.generate_installers import generate_installers
 
 from superflore.generators.bitbake.gen_packages import regenerate_installer
-from superflore.generators.bitbake.ros_meta import ros_meta
+from superflore.generators.bitbake.ros_meta import RosMeta
 
 from superflore.utils import err
 from superflore.utils import info
@@ -38,39 +36,27 @@ preserve_existing = True
 overlay = None
 
 
-def link_existing_files(mode):
+def clean_up(distro, preserve_repo=False):
     global overlay
-    sym_link_msg = 'Symbolicly linking files from {0}/recipes-ros-{1}...'
-    dir_fmt = '{0}/recipes-ros-{1}'
-    if mode == 'all' or mode == 'update':
-        for x in active_distros:
-            info(sym_link_msg.format(overlay.repo_dir, x))
-            os.symlink(
-                dir_fmt.format(overlay.repo_dir, x), './recipes-ros-' + x
-            )
-    else:
-        # only link the relevant directory.
-        info(sym_link_msg.format(overlay.repo_dir, mode))
-        os.symlink(
-            dir_fmt.format(overlay.repo_dir, mode), './recipes-ros-' + mode
-        )
+    if not preserve_repo:
+        clean_msg = \
+            'Cleaning up tmp directory {0}...'.format(overlay.repo.repo_dir)
+        info(clean_msg)
+        shutil.rmtree(overlay.repo_dir)
 
 
-def clean_up(distro):
-    global overlay
-    clean_msg = 'Cleaning up tmp directory {0}...'.format(overlay.repo_dir)
-    info(clean_msg)
-    shutil.rmtree(overlay.repo_dir)
-    info('Cleaning up symbolic links...')
-    if mode != 'all' and mode != 'update':
-        os.remove('recipes-ros-{0}'.format(distro))
-    else:
-        for x in active_distros:
-            os.remove('recipes-ros-{0}'.format(x))
+def file_pr(overlay, delta, missing_deps):
+    try:
+        overlay.pull_request('{0}\n{1}'.format(delta, missing_deps))
+    except Exception as e:
+        err('Failed to file PR with allenh1/meta-ros repo!')
+        err('Exception: {0}'.format(e))
+        sys.exit(1)
 
 
 def main():
     global overlay
+    global preserve_existing
 
     parser = argparse.ArgumentParser('Deploy ROS packages into Yocto Linux')
     parser.add_argument(
@@ -83,32 +69,25 @@ def main():
         help='regenerate all packages in all distros',
         action="store_true"
     )
+    parser.add_argument(
+        '--output-repository-path',
+        help='location of the Git repo',
+        type=str
+    )
+
     args = parser.parse_args(sys.argv[1:])
     # clone current repo
-    overlay = ros_meta()
+    overlay = RosMeta(args.output_repository_path)
     selected_targets = active_distros
 
     if args.all:
         warn('"All" mode detected... this may take a while!')
+        preserve_existing = False
     elif args.ros_distro:
         selected_targets = [args.ros_distro]
         preserve_existing = False
-
     # clone current repo
     selected_targets = active_distros
-    for x in active_distros:
-        try:
-            os.remove('recipes-ros-{0}'.format(x))
-            warn_msg =\
-                'removing existing symlink "./recipes-ros-{0}"'.format(x)
-            warn(warn_msg)
-        except OSError as e:
-            if e.errno == errno.ENOENT:
-                pass
-            else:
-                raise e
-    link_existing_files(args.ros_distro)
-
     # generate installers
     total_installers = dict()
     total_broken = set()
@@ -187,5 +166,5 @@ def main():
         err('Exception: {0}'.format(e))
         sys.exit(1)
 
-    clean_up()
+    clean_up(args.ros_distro, args.output_repository_path)
     ok('Successfully synchronized repositories!')
